app/requirements.txt
Django==4.1.2
djangorestframework
django-filter
django-cors-headers
django-import-export
djangorestframework-simplejwt
dj-rest-auth
gunicorn
Pillow
psycopg2
psycopg2-binary
mysqlclient
kavenegar